now use Maven project changed listener
